Move MenuFeatureSelectionControllers to menu packages.
